+++ src/routing/mod.rs
@@ -425,11 +425,8 @@ pub trait RoutingDsl: crate::sealed::Sealed + Sized {
     /// Check that your service cannot fail.
     ///
     /// That is, its error type is [`Infallible`].
-    fn check_infallible<ReqBody>(self) -> Self
-    where
-        Self: Service<Request<ReqBody>, Error = Infallible>,
-    {
-        self
+    fn check_infallible(self) -> CheckInfallible<Self> {
+        CheckInfallible(self)
     }
 }
 
@@ -918,6 +915,35 @@ fn strip_prefix(uri: &Uri, prefix: &str) -> Uri {
 
     Uri::from_parts(parts).unwrap()
 }
+
+/// Middleware that statically verifies that a service cannot fail.
+///
+/// Created with [`check_infallible`](RoutingDsl::check_infallible).
+#[derive(Debug, Clone, Copy)]
+pub struct CheckInfallible<S>(S);
+
+impl<R, S> Service<R> for CheckInfallible<S>
+where
+    S: Service<R, Error = Infallible>,
+{
+    type Response = S::Response;
+    type Error = S::Error;
+    type Future = S::Future;
+
+    #[inline]
+    fn poll_ready(&mut self, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        self.0.poll_ready(cx)
+    }
+
+    #[inline]
+    fn call(&mut self, req: R) -> Self::Future {
+        self.0.call(req)
+    }
+}
+
+impl<S> RoutingDsl for CheckInfallible<S> {}
+
+impl<S> crate::sealed::Sealed for CheckInfallible<S> {}
 
 #[cfg(test)]
 mod tests {